[BUD-41] refactor : 다이어리 캘린더 뷰 보이지않던  픽스(#25)

+++ diary/src/main/java/com/nocapstone/diary/dto/DiaryData.kt
@@ -6,10 +6,10 @@ data class DiaryData(
 
 data class Diary(
     val day : Int,
-    val diaryId: Int,
+    val diaryId: Long,
     val thumbnail: String?
 )
 
 data class DiaryId(
-    val diaryId: Int
+    val diaryId: Long
 )
